Fixed bug where the map URL wouldn't be launched first time
The `application:openURL:sourceApplication:annotation` message is sent
to an app that is already running. If the app is *launched* with a url,
then the url will be in the launchOptions dictionary passed in to
`application:didFinishLaunchingWithOptions:` and the `openURL` message
is not sent to the app delegate for that url.

This patch fixes that by checking for the launch url and passing it
through to the `openURL` message. I needed to wait for 0.3 seconds for
the system to "settle down" after launch before triggering the url,
otherwise, nothing would happen.